+++ resources/practicalli/service/dev/mulog_events.clj
@@ -1,22 +1,16 @@
 ;; ---------------------------------------------------------
 ;; Mulog Global Context and Custom Publisher
 ;;
-;; - set event log global context
 ;; - tap publisher for use with Portal and other tap sources
 ;; - publish all mulog events to Portal tap source
+;;
+;; NOTE: Mulog global context set in system configuration
 ;; ---------------------------------------------------------
 
 (ns mulog-events
   (:require
    [com.brunobonacci.mulog        :as mulog]
    [com.brunobonacci.mulog.buffer :as mulog-buffer]))
-
-;; ---------------------------------------------------------
-;; Set event global context
-;; - information added to every event for REPL workflow
-(mulog/set-global-context! {:app-name "{{main/ns}} Service",
-                            :version "0.1.0", :env "dev"})
-;; ---------------------------------------------------------
 
 ;; ---------------------------------------------------------
 ;; Mulog event publishing